chore(ci): export PYTHONPATH from setup-toolchain action
The per-package ``test-unit / unit-tests (gpg-bridge)`` shard
introduced in #491 fails 19 of 95 tests because the shard runs
raw ``pytest`` without sourcing ``scripts/_bootstrap_venv.sh``,
which is the only place the workspace exports
``packages/<pkg>/tests`` on ``PYTHONPATH`` so subprocesses
launched by tests (``python -m gpg_backend_fake``,
``python -m things_client_fake``) can resolve the test-only
fakes. Pytest's ``pythonpath`` config only affects the in-process
``sys.path``, not subprocess ``PYTHONPATH``.

Move the export into ``.github/actions/setup-toolchain`` so every
workflow that uses the action (per-package shards, verify-standards,
codeql, ...) inherits the right environment from $GITHUB_ENV
without sourcing the bootstrap script. Mirrors the bootstrap export
verbatim so local + CI behaviour stays in lockstep; the bootstrap
script is unchanged so local developers keep their existing flow.

With the action-side export in place, the #520 / #524 workaround
that relocated the gpg-subprocess + HTTP wiring tests under
``packages/gpg-bridge/tests/integration/`` is no longer needed.
Revert it end-to-end:

- ``git mv`` the 5 relocated tests back to
  ``packages/gpg-bridge/tests/`` (test_gpg_bridge_server,
  test_gpg_bridge_perf_budget,
  test_gpg_subprocess_client[_passphrase] back to ``tests/``;
  test_bridge_network_drop back to ``tests/fault/``).
- Re-fold the gpg_backend_fake-driven wedge case from
  ``tests/integration/test_wedged_gpg_subprocess.py`` back into
  ``tests/fault/test_backend_subprocess_hang.py`` and delete the
  extracted file.
- Restore ``--cov-fail-under=62`` in
  ``packages/gpg-bridge/pyproject.toml`` (was lowered to 55 as
  the workaround).
- Restore the ``test_gpg_bridge_server`` / ``test_gpg_subprocess_client``
  entries in the workspace ``[mypy.overrides] module`` and the
  ``test_gpg_bridge_server.py`` entry in ``[mutmut] tests_dir``,
  both dropped when the files moved out of ``tests/``.
- Drop the ``# integration-isolation: in-process-server`` carve-out
  from ``scripts/verify-integration-isolation.sh``; with no
  in-process tests left under ``tests/integration/`` the marker
  no longer has any callers and the simpler grep loop is the
  pre-#524 default.

Verified locally: ``pytest packages/gpg-bridge/tests/
--ignore=packages/gpg-bridge/tests/integration --no-cov`` runs
95 passing tests with the action-style PYTHONPATH set, and 19
failing without it (same shape as the original CI failure mode).

Supersedes #523 — the codeql-config.yml and ADR 0040 path
references it tracked become valid again once the tests are
relocated back; #523 has been closed.

Closes: #522
