+++ packages/gpg-bridge/tests/fault/test_bridge_network_drop.py
@@ -1,7 +1,6 @@
 # SPDX-FileCopyrightText: 2026 Aidan Nagorcka-Smith
 #
 # SPDX-License-Identifier: MIT
-# integration-isolation: in-process-server
 
 """Fault-injection: client drops the TCP connection mid-request.
