+++ packages/gpg-bridge/tests/test_gpg_bridge_perf_budget.py
@@ -1,7 +1,6 @@
 # SPDX-FileCopyrightText: 2026 Aidan Nagorcka-Smith
 #
 # SPDX-License-Identifier: MIT
-# integration-isolation: in-process-server
 
 """Performance-budget assertion for gpg-bridge sign / verify.
