+++ pyproject.toml
@@ -327,10 +327,12 @@ module = [
     "test_gpg_backend_fake",
     "test_gpg_bridge_cli",
     "test_gpg_bridge_config",
+    "test_gpg_bridge_server",
     "test_gpg_cli_cli",
     "test_gpg_cli_client",
     "test_gpg_cli_config",
     "test_gpg_models",
+    "test_gpg_subprocess_client",
     "test_integration_harness",
     "test_integration_support",
     "test_keyring_unavailable",
@@ -443,6 +445,7 @@ tests_dir = [
     "packages/agent-auth/tests/test_tokens.py",
     "packages/gpg-bridge/tests/test_gpg_backend_fake.py",
     "packages/gpg-bridge/tests/test_gpg_bridge_config.py",
+    "packages/gpg-bridge/tests/test_gpg_bridge_server.py",
     "packages/gpg-cli/tests/test_gpg_cli_cli.py",
     "packages/gpg-cli/tests/test_gpg_cli_config.py",
     # keep-sorted end
